% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{control}
\alias{control}
\title{Suspend or resume profiling. Actually this is just wrapper on (undocumented) function moncontrol}
\usage{
control(v)
}
\arguments{
\item{v}{0 to suspend, 1 to resume}
}
\description{
Suspend or resume profiling. Actually this is just wrapper on (undocumented) function moncontrol
}
