% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lsmod}
\alias{lsmod}
\title{get list of loaded modules inside current process}
\usage{
lsmod()
}
\value{
string vector of modules with full paths
}
\description{
get list of loaded modules inside current process
}
