% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lsmodf}
\alias{lsmodf}
\title{get frame of loaded modules inside current process}
\usage{
lsmodf()
}
\value{
frame consisted of modules full paths, base address and size
}
\description{
get frame of loaded modules inside current process
}
