% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lsmodfr}
\alias{lsmodfr}
\title{get frame of loaded modules inside current process matching regex}
\usage{
lsmodfr(re)
}
\arguments{
\item{re}{regex}
}
\value{
frame consisted of modules full paths, base address and size
}
\description{
get frame of loaded modules inside current process matching regex
}
