% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prof_start}
\alias{prof_start}
\title{Start profiling of some module}
\usage{
prof_start(fname)
}
\arguments{
\item{fname}{full path to module}
}
\value{
-1 if no such module or it is not ELF, 0 when module cannot be profiles, 1 if it was conpuled with -pg option, 2 fir -finstrument-functions option
}
\description{
Start profiling of some module
}
