% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prof_stop}
\alias{prof_stop}
\title{Stop profiling. results will be saved in current directory in file gmon.base_address_of_module.PID}
\usage{
prof_stop()
}
\value{
0 when no profiling happened, 1 otherwise
}
\description{
Stop profiling. results will be saved in current directory in file gmon.base_address_of_module.PID
}
